Reverted revision(s) 53286 from branches/FeatureBranches/nghs-1d2dflooding: FM1D2D-1101 Make sure global precipitation data is correctly written. ........
Former-commit-id: fd3adf93468f73b32a7f1fbc5b2ccd23e2d4fead
Former-commit-id: 603bd78c1486a81c6254546dd444abd6ad216ff3

+++ src/DeltaShell.Plugins.DelftModels.RainfallRunoff/Exporters/RainfallRunoffModelExporter.cs
@@ -73,9 +73,12 @@ namespace DeltaShell.Plugins.DelftModels.RainfallRunoff.Exporters
                 meteoWriter.Export(model.Evaporation, evapFile);
             }
 
-            var precipitationFile = Path.Combine(Path.GetFullPath(path), "default.bui");
-            FileUtils.DeleteIfExists(precipitationFile);
-            meteoWriter.Export(model.Precipitation, precipitationFile);
+            if (model.Precipitation.DataDistributionType != MeteoDataDistributionType.Global)
+            {
+                var precipitationFile = Path.Combine(Path.GetFullPath(path), "default.bui");
+                FileUtils.DeleteIfExists(precipitationFile);
+                meteoWriter.Export(model.Precipitation, precipitationFile);
+            }
 
             if (model.Temperature.DataDistributionType != MeteoDataDistributionType.Global)
             {
